refactor: 태그 만드는 함수 분리
